Refactor and document PathAware visitor classes
1. Renamed PathAwareAbstractVisitor to PathAwareVisitor (removing the
interface) and added the default (no-op) method impls from
PathAwareVisitorAdapter (allowing us to delete that one).
2. Introduced a reset() method to enable re-using a visitor instance
even after an exception has occurred
3. Add javadoc to explain how to use this class and especially its
public getPath and reset methods.
