+++ fhir-model/src/test/java/com/ibm/fhir/model/path/test/FHIRPathEvaluatorTest.java
@@ -27,7 +27,7 @@ import com.ibm.fhir.model.type.Instant;
 import com.ibm.fhir.model.type.Integer;
 import com.ibm.fhir.model.type.Meta;
 import com.ibm.fhir.model.type.String;
-import com.ibm.fhir.model.visitor.PathAwareAbstractVisitor;
+import com.ibm.fhir.model.visitor.PathAwareVisitor;
 
 public class FHIRPathEvaluatorTest {
     public static void main(java.lang.String[] args) throws Exception {
@@ -83,7 +83,7 @@ public class FHIRPathEvaluatorTest {
         
         System.out.println("");
         
-        PathAwareAbstractVisitor.DEBUG = true;
+        PathAwareVisitor.DEBUG = true;
         
         FHIRPathEvaluator evaluator = FHIRPathEvaluator.evaluator();
         EvaluationContext evaluationContext = new EvaluationContext(patient);
