+++ fhir-persistence-jdbc/src/test/java/com/ibm/fhir/persistence/jdbc/test/spec/BaseOperation.java
@@ -7,7 +7,7 @@
 package com.ibm.fhir.persistence.jdbc.test.spec;
 
 import com.ibm.fhir.model.resource.Resource;
-import com.ibm.fhir.model.spec.test.ResourceComparatorVisitor;
+import com.ibm.fhir.model.util.test.ResourceComparatorVisitor;
 import com.ibm.fhir.persistence.exception.FHIRPersistenceException;
 import com.ibm.fhir.persistence.util.ResourceFingerprintVisitor;
 import com.ibm.fhir.persistence.util.SaltHash;
