+++ fhir-search/src/main/java/com/ibm/fhir/search/compartment/CompartmentUtil.java
@@ -39,16 +39,17 @@ import com.ibm.fhir.search.exception.SearchExceptionUtil;
  * The compartments are defined using FHIR R4 CompartmentDefinitions. <br>
  * The R4 CompartmentDefintions and boundaries are defined at https://www.hl7.org/fhir/compartmentdefinition.html <br>
  * <br>
- * CompartmentDefintion: <br>
+ * CompartmentDefintion:
+ * <ul>
  * <li>Patient - https://www.hl7.org/fhir/compartmentdefinition-patient.json</li>
  * <li>Encounter - https://www.hl7.org/fhir/compartmentdefinition-encounter.json</li>
  * <li>RelatedPerson - https://www.hl7.org/fhir/compartmentdefinition-relatedperson.json</li>
- * <li>Practioner - https://www.hl7.org/fhir/compartmentdefinition-practitioner.json</li>
- * <li>Device - https://www.hl7.org/fhir/compartmentdefinition-device.json</li> <br>
+ * <li>Practitioner - https://www.hl7.org/fhir/compartmentdefinition-practitioner.json</li>
+ * <li>Device - https://www.hl7.org/fhir/compartmentdefinition-device.json</li>
+ * </ul>
  * 
- * This class extracts the Compartment Logic from SearchUtil, converted the Custom Compartment logic and format into the
- * ComponentDefintion, adds support for the the default definition, and refactors the code to create constants and limit
- * extraneous strings and duplicate definitions<br/>
+ * This class extracts the Compartment Logic from SearchUtil, converts the Custom Compartment logic and format into the
+ * ComponentDefintion, adds support for the the default definition.
  * 
  * <br>
  * Load the class in the classloader to initialize static members. Call this before using the class in order to avoid a
