+++ fhir-validation/src/test/java/com/ibm/fhir/validation/test/BundleValidationTest.java
@@ -36,7 +36,7 @@ import com.ibm.fhir.model.type.Uri;
 import com.ibm.fhir.model.type.code.BundleType;
 import com.ibm.fhir.model.type.code.HTTPVerb;
 import com.ibm.fhir.model.type.code.IssueSeverity;
-import com.ibm.fhir.model.visitor.PathAwareAbstractVisitor;
+import com.ibm.fhir.model.visitor.PathAwareVisitor;
 import com.ibm.fhir.validation.FHIRValidator;
 
 public class BundleValidationTest {
@@ -103,7 +103,7 @@ public class BundleValidationTest {
         FHIRGenerator.generator(Format.JSON, true).generate(bundle, System.out);
         System.out.println();
         
-        PathAwareAbstractVisitor.DEBUG = false;
+        PathAwareVisitor.DEBUG = false;
         FHIRValidator.DEBUG = false;
         List<Issue> issues = FHIRValidator.validator().validate(bundle);
         
@@ -145,7 +145,7 @@ public class BundleValidationTest {
         FHIRGenerator.generator(Format.JSON, true).generate(bundle, System.out);
         System.out.println();
         
-        PathAwareAbstractVisitor.DEBUG = false;
+        PathAwareVisitor.DEBUG = false;
         FHIRValidator.DEBUG = false;
         List<Issue> issues = FHIRValidator.validator().validate(bundle);
         
@@ -200,7 +200,7 @@ public class BundleValidationTest {
         FHIRGenerator.generator(Format.JSON, true).generate(patient, System.out);
         System.out.println();
         
-        PathAwareAbstractVisitor.DEBUG = false;
+        PathAwareVisitor.DEBUG = false;
         FHIRValidator.DEBUG = false;
         List<Issue> issues = FHIRValidator.validator().validate(bundle);
         
@@ -232,7 +232,7 @@ public class BundleValidationTest {
         FHIRGenerator.generator(Format.JSON, true).generate(patient, System.out);
         System.out.println();
         
-        PathAwareAbstractVisitor.DEBUG = false;
+        PathAwareVisitor.DEBUG = false;
         FHIRValidator.DEBUG = false;
         List<Issue> issues = FHIRValidator.validator().validate(bundle);
