+++ fhir-validation/src/test/java/com/ibm/fhir/validation/test/FHIRValidatorTest.java
@@ -30,7 +30,7 @@ import com.ibm.fhir.model.type.Integer;
 import com.ibm.fhir.model.type.Meta;
 import com.ibm.fhir.model.type.String;
 import com.ibm.fhir.model.type.code.IssueSeverity;
-import com.ibm.fhir.model.visitor.PathAwareAbstractVisitor;
+import com.ibm.fhir.model.visitor.PathAwareVisitor;
 import com.ibm.fhir.validation.FHIRValidator;
 
 public class FHIRValidatorTest {
@@ -82,7 +82,7 @@ public class FHIRValidatorTest {
         
         System.out.println("");
         
-        PathAwareAbstractVisitor.DEBUG = false;
+        PathAwareVisitor.DEBUG = false;
         FHIRValidator.DEBUG = false;
         List<Issue> issues = FHIRValidator.validator().validate(patient);
